refactor(wallet): Adding global management for Stellar Wallets
