+++ apps/web-app/src/features/wallet/components/WalletButton.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from "react";
 import { ConnectButton } from "@rainbow-me/rainbowkit";
 import { useDisconnect } from "wagmi";
-import { useWalletConnection } from "@/hooks/useWalletConnection";
+import { useStellarWallet } from "@/hooks/useStellarWallet";
 import { useWalletType } from "@/hooks/useWalletType";
 import { ConnectWalletModal } from "./ConnectWalletModal";
 
@@ -20,7 +20,7 @@ export const WalletButton: React.FC = () => {
     evmAddress,
     stellarAddress,
   } = useWalletType();
-  const { disconnect: disconnectStellar } = useWalletConnection();
+  const { disconnect: disconnectStellar } = useStellarWallet();
   const { disconnect: disconnectEvm } = useDisconnect();
   const [showModal, setShowModal] = useState(false);
   const address = walletType === "evm" ? evmAddress : stellarAddress;
