+++ apps/web-app/src/hooks/useWalletType.ts
@@ -2,16 +2,13 @@ import { useAccount } from "wagmi";
 import { useWallet as useStellarWallet } from "./useWallet";
 
 /**
- * Hook to detect which type of wallet is connected (EVM or Stellar)
- * Returns the wallet type and connection status
+ * Hook to detect which type of wallet is connected (EVM or Stellar).
+ * Returns the wallet type and connection status.
  */
 export type WalletType = "evm" | "stellar" | "none";
 
 export const useWalletType = () => {
-  // Check EVM wallet connection
   const { isConnected: isEvmConnected, address: evmAddress } = useAccount();
-
-  // Check Stellar wallet connection
   const { address: stellarAddress } = useStellarWallet();
 
   // Determine wallet type
