+++ apps/web-app/src/lib/helpers/wallet.ts
@@ -1,70 +1,18 @@
-"use client";
-
-import {
-  StellarWalletsKit,
-  FREIGHTER_ID,
-  FreighterModule,
-  AlbedoModule,
-  xBullModule,
-  LobstrModule,
-  WalletNetwork,
-} from "@creit.tech/stellar-wallets-kit";
-import { LedgerModule } from "@creit.tech/stellar-wallets-kit/modules/ledger.module";
-import {
-  WalletConnectAllowedMethods,
-  WalletConnectModule,
-} from "@creit.tech/stellar-wallets-kit/modules/walletconnect.module";
 import { Horizon } from "@stellar/stellar-sdk";
-import {
-  networkPassphrase,
-  stellarNetwork,
-  horizonUrl,
-} from "../constants/network";
+import { stellarNetwork, horizonUrl } from "../constants/network";
+import { getStellarWalletKit } from "./stellar-wallet-kit";
 
-let stellarWalletKitInstance: StellarWalletsKit | null = null;
+/** Reexport del Kit para compatibilidad; usar getStellarWalletKit() desde helpers. */
+export const getWallet = () => getStellarWalletKit();
 
-const getKit = (): StellarWalletsKit => {
-  if (typeof window === "undefined") {
-    throw new Error("StellarWalletsKit can only be used in the browser");
-  }
-
-  if (!stellarWalletKitInstance) {
-    stellarWalletKitInstance = new StellarWalletsKit({
-      network: networkPassphrase as WalletNetwork,
-      selectedWalletId: FREIGHTER_ID,
-      modules: [
-        new FreighterModule(),
-        new AlbedoModule(),
-        new xBullModule(),
-        new LedgerModule(),
-        new LobstrModule(),
-        new WalletConnectModule({
-          url: "https://nekoprotocol.xyz",
-          projectId:
-            process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID ||
-            "fa57d523d12455e4fc2c8c83c94ec7b1",
-          method: WalletConnectAllowedMethods.SIGN,
-          name: "Neko Protocol",
-          description: "Neko — The Marketplace for Real-World Assets On-Chain",
-          icons: ["/Neko.svg"],
-          network: networkPassphrase as WalletNetwork,
-        }),
-      ],
-    });
-  }
-
-  return stellarWalletKitInstance;
-};
-
+// Create Horizon server instance lazily (only when needed and in browser)
 const getHorizon = (): Horizon.Server | null => {
   if (typeof window === "undefined") {
     return null;
   }
-
   if (!horizonUrl) {
     return null;
   }
-
   try {
     return new Horizon.Server(horizonUrl, {
       allowHttp: stellarNetwork === "LOCAL",
@@ -82,12 +30,10 @@ export const fetchBalances = async (address: string) => {
   if (typeof window === "undefined") {
     return {};
   }
-
   const horizonInstance = getHorizon();
   if (!horizonInstance) {
     return {};
   }
-
   try {
     const { balances } = await horizonInstance
       .accounts()
@@ -95,10 +41,7 @@ export const fetchBalances = async (address: string) => {
       .call();
     const mapped = balances.reduce((acc, b) => {
       const formattedBalance = formatter.format(Number(b.balance));
-      const balanceEntry = {
-        ...b,
-        balance: formattedBalance,
-      };
+      const balanceEntry = { ...b, balance: formattedBalance };
       const key =
         b.asset_type === "native"
           ? "xlm"
@@ -113,5 +56,3 @@ export const fetchBalances = async (address: string) => {
     return {};
   }
 };
-
-export const getWallet = () => getKit();
